Move version class
